requirements.txt
asgiref==3.8.1
astroid==3.2.4
asttokens==2.4.1
async-timeout==4.0.3
babel==2.16.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
black==24.8.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
decorator==5.1.1
dill==0.3.9
distlib==0.3.8
Django==4.2.16
django-appconf==1.0.6
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-extra-views==0.14.0
django-haystack==3.3.0
-e git+https://github.com/django-oscar/django-oscar.git@bd92e7208cfe503026b61b762ec58606c942f027#egg=django_oscar
django-phonenumber-field==8.0.0
django-redis==5.3.0
django-select2==8.2.1
django-tables2==2.7.0
django-treebeard==4.7.1
django-webtest==1.9.12
django-widget-tweaks==1.5.0
easy-thumbnails==2.9
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
factory_boy==3.3.1
Faker==30.0.0
filelock==3.16.1
freezegun==1.5.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mock==5.1.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.46
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
purl==1.6
pycountry==24.6.1
Pygments==2.18.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyprof2calltree==1.4.5
pyproject-api==1.8.0
pysolr==3.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.2
redis==5.0.8
requests==2.32.3
six==1.16.0
sorl-thumbnail==12.11.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
straight.plugin==1.5.0
tomli==2.0.1
tomlkit==0.13.2
tox==4.20.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uWSGI==2.0.27
vdt.version==0.1.8
vdt.versionplugin.default==0.0.3
vdt.versionplugin.wheel==0.0.8
virtualenv==20.26.6
waitress==3.0.0
wcwidth==0.2.13
WebOb==1.8.8
WebTest==3.0.1
whitenoise==6.7.0
Whoosh==2.7.4
zipp==3.20.2
openpyxl
